be/sim DB 테이블 도메인 생성, UserProfiles 관련 로직 주석처리, sql 쿼리 업데이트

+++ src/main/java/teammate/teammate/repository/UserRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class UserRepository {
     private final EntityManager em;
 
-    public Users getUserByUsername(BigInteger user_id){
+    public Users getUserByUsername(int user_id){
         String query = "select u from Users u";
 
         List<Users> resultList = em.createQuery(query, Users.class).getResultList();
